Move delay in `test_utils` as it's only used in tests
