+++ src/test_utils/async_fs.ts
@@ -8,7 +8,8 @@
 import fs, { promises as asyncFs } from "fs";
 import { promisify } from "util";
 
-import { delay } from "../lib/delay";
+import { delay } from "./delay";
+
 export const existsAsync = (filepath: string): Promise<void> =>
   asyncFs.access(filepath, fs.constants.F_OK);
